+++ src/lib/components/Access/ProtectionButtons.js
@@ -7,51 +7,49 @@
 // under the terms of the MIT License; see LICENSE file for more details.
 import React, { Component } from 'react';
 import { Button } from 'semantic-ui-react';
-import { FastField } from 'formik';
+import { Field } from 'formik';
 import { i18next } from "@translations/invenio_app_rdm/i18next";
+import PropTypes from 'prop-types';
 
 class ProtectionButtonsComponent extends Component {
-  /**
-   * Returns the props for a protection button.
-   * @param active is button active
-   * @param activeColor button color when active
-   */
-  getButtonProps(active, activeColor) {
-    let props = { active };
-    if (active) {
-      props['color'] = activeColor;
-    }
-    return props;
-  }
+  handlePublicButtonClick = (event, elemProps) => {
+    const { formik, fieldPath } = this.props;
+    formik.form.setFieldValue(fieldPath, 'public');
+    // NOTE: We reset values, so if embargo filled and click Public,
+    //       user needs to fill embargo again. Otherwise, lots of
+    //       bookkeeping.
+    formik.form.setFieldValue('access.embargo', {
+      active: false,
+    });
+  };
+
+  handleRestrictionButtonClick = (event, elemProps) => {
+    const { formik, fieldPath } = this.props;
+    formik.form.setFieldValue(fieldPath, 'restricted');
+  };
 
   render() {
-    const { fieldPath, formik, active } = this.props;
+    const { active, disabled } = this.props;
+
+    const publicColor = active ? 'positive' : '';
+    const restrictedColor = !active ? 'negative' : '';
 
     return (
-      <Button.Group widths={'2'}>
+      <Button.Group widths="2">
         <Button
-          {...this.getButtonProps(active, 'green')}
-          onClick={(event, data) => {
-            formik.form.setFieldValue(fieldPath, 'public');
-            // NOTE: We reset values, so if embargo filled and click Public,
-            //       user needs to fill embargo again. Otherwise lots of
-            //       bookkeeping.
-            formik.form.setFieldValue('access.embargo', {
-              active: false,
-            });
-          }}
-          compact
-          attached
+          className={publicColor}
+          data-testid="protection-buttons-component-public"
+          disabled={disabled}
+          onClick={this.handlePublicButtonClick}
+          active={active}
         >
           {i18next.t('Public')}
         </Button>
         <Button
-          {...this.getButtonProps(!active, 'red')}
-          onClick={(event, data) =>
-            formik.form.setFieldValue(fieldPath, 'restricted')
-          }
-          compact
-          attached
+          className={restrictedColor}
+          data-testid="protection-buttons-component-restricted"
+          active={!active}
+          onClick={this.handleRestrictionButtonClick}
         >
           {i18next.t('Restricted')}
         </Button>
@@ -60,12 +58,24 @@ class ProtectionButtonsComponent extends Component {
   }
 }
 
+ProtectionButtonsComponent.propTypes = {
+  fieldPath: PropTypes.string.isRequired,
+  formik: PropTypes.object.isRequired,
+  active: PropTypes.bool,
+  disabled: PropTypes.bool,
+};
+
+ProtectionButtonsComponent.defaultProps = {
+  active: true,
+  disabled: false,
+};
+
 export class ProtectionButtons extends Component {
   render() {
     const { fieldPath } = this.props;
 
     return (
-      <FastField
+      <Field
         name={fieldPath}
         component={(formikProps) => (
           <ProtectionButtonsComponent formik={formikProps} {...this.props} />