+++ src/lib/components/FormFeedback.js
@@ -48,7 +48,6 @@ const defaultLabels = {
   'metadata.identifiers': i18next.t('Alternate identifiers'),
   'access.embargo.until': i18next.t('Embargo until'),
   'pids.doi': i18next.t('DOI'),
-  pids: i18next.t('PIDS'),
 };
 
 const ACTIONS = {
@@ -233,10 +232,12 @@ class DisconnectedFormFeedback extends Component {
 
     return labelledErrorMessages;
   }
+
   render() {
     const actionState = this.props.actionState;
 
     const errors = this.props.errors || {};
+
     const { feedback, message } = _get(ACTIONS, actionState, {
       feedback: undefined,
       message: undefined,
@@ -256,6 +257,9 @@ class DisconnectedFormFeedback extends Component {
       )
     );
 
+    // errors not related to validation, following a different format {status:.., message:..}
+    const backendErrorMessage = errors.message;
+
     return (
       <Message
         visible
@@ -266,7 +270,7 @@ class DisconnectedFormFeedback extends Component {
       >
         <Grid container>
           <Grid.Column width={15} textAlign="left">
-            <strong>{message}</strong>
+            <strong>{backendErrorMessage || message}</strong>
             {listErrors.length > 0 && <Message.List>{listErrors}</Message.List>}
           </Grid.Column>
         </Grid>