+++ src/lib/components/RelatedWorksField.js
@@ -90,6 +90,7 @@ export class RelatedWorksField extends Component {
 
                 <Form.Field>
                   <Button
+                    aria-label={i18next.t("Remove field")}
                     className="close-btn"
                     icon 
                     onClick={() => arrayHelpers.remove(indexPath)}