+++ src/lib/components/SaveButton.js
@@ -8,22 +8,26 @@
 
 import { i18next } from "@translations/invenio_app_rdm/i18next";
 import React, { Component } from 'react';
-import { ActionButton } from 'react-invenio-forms';
 import { connect } from 'react-redux';
-import { Icon } from 'semantic-ui-react';
+import { Button } from 'semantic-ui-react';
 import {
   DepositFormSubmitActions,
   DepositFormSubmitContext,
 } from '../DepositFormSubmitContext';
 import { DRAFT_SAVE_STARTED } from '../state/types';
 import { scrollTop } from '../utils';
+import _omit from 'lodash/omit';
+import { connect as connectFormik } from 'formik';
 
 export class SaveButtonComponent extends Component {
   static contextType = DepositFormSubmitContext;
 
-  handleSave = (event, formik) => {
+  handleSave = (event) => {
+    const { formik } = this.props;
+    const { handleSubmit } = formik;
+
     this.context.setSubmitContext(DepositFormSubmitActions.SAVE);
-    formik.handleSubmit(event);
+    handleSubmit(event);
     // MSD-LIVE CHANGE added timeout
     //I'm not sure why this has to be in a timeout but it works as long as it is SIGH
     setTimeout(function(){
@@ -32,28 +36,22 @@ export class SaveButtonComponent extends Component {
   };
 
   render() {
-    const { actionState, ...uiProps } = this.props;
+    const { actionState, formik, ...ui } = this.props;
+    const { isSubmitting } = formik;
+
+    const uiProps = _omit(ui, ['dispatch']);
 
     return (
-      <ActionButton
-        isDisabled={(formik) => formik.isSubmitting}
+      <Button
         name="save"
-        onClick={this.handleSave}
-        icon
+        disabled={isSubmitting}
+        onClick={(event) => this.handleSave(event)}
+        icon="save"
+        loading={isSubmitting && actionState === DRAFT_SAVE_STARTED}
         labelPosition="left"
+        content={i18next.t('Save draft')}
         {...uiProps}
-      >
-        {(formik) => (
-          <>
-            {formik.isSubmitting && actionState === DRAFT_SAVE_STARTED ? (
-              <Icon size="large" loading name="spinner" />
-            ) : (
-              <Icon name="save" />
-            )}
-            {i18next.t('Save draft')}
-          </>
-        )}
-      </ActionButton>
+      />
     );
   }
 }
@@ -62,4 +60,7 @@ const mapStateToProps = (state) => ({
   actionState: state.deposit.actionState,
 });
 
-export const SaveButton = connect(mapStateToProps, null)(SaveButtonComponent);
+export const SaveButton = connect(
+  mapStateToProps,
+  null
+)(connectFormik(SaveButtonComponent));